requirements.txt
streamlit>=1.35
msal>=1.28
semantic-kernel>=1.20.0
requests>=2.31
python-dotenv>=1.0
pytz>=2024
nest_asyncio>=1.6
